Correct (hopefully) fix for autopm get/put
So it appears the problem was that usb_autopm_get_interface() will wake
up sleeping hardware, which isn't being allowed when called in this
context for some reason (maybe because it required process context or
maybe something else). However, we don't need this because there's no
way the device could have even be put to sleep yet, so I've replaced it
with usb_autopm_get_interface_no_resume(), which just increments the
counter to make sure it doesn't go to sleep until we've completed the
mcp2210_configure().