Work-around for usb_autopm_get_interface() returning non-zero
Unfortunately, the contract for usb_autopm_get_interface() is not
spelled out with proper kerneldocs. However, we should assume that it
will return non-zero in case of failure and zero in all other cases.
Sadly, this contract is violated due to a change that allowed you to
know when you didn't need to actually restart anything (in the PM layer)
because nothing was asleep. This change accounts for this bug.

This should correct the remaining problems in
https://github.com/daniel-santos/mcp2210-linux/issues/5