Modify core, state and IRQ struct fields.
First, we've had fields in mcp2210_state that do not belong there. This
class represents the state of the hardware, and not the linux device
(representation and usage of that hardware).  Fields that are releated
to hardware state stay there and fields that are part of the Linux
device state go into mcp2210_device.

Finally, we modify important fields for IRQ support.