mcp2210-spi.c: BUG_ON NULL tx_buf, tolerate NULL rx_buf
First this fixes a logical flaw that fails to verify that the tx_buf of
a struct spi_transfer is non-null.  The check:

	if (!(xfer->tx_buf || xfer->rx_buf))

is only true if BOTH of the buffers are NULL.

Second, the driver should probably tollerate a situation where the
rx_buf is NULL, so this fixes that for cases where CONFIG_MCP2210_DEBUG
is enabled.

This fixes issue #35.